fix: change price to Decimal

+++ back-end/src/database/migrations/20220504180502-create-sales.js
@@ -29,7 +29,7 @@ module.exports = {
         },
       },
       total_price: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(9,2),
         allowNull: false,
 
       },
